Since I'm onto the backend and keeping it in the same directory, renamed 'frontend' dirs to 'compiler'
