+++ include/compiler/datatypes/partials.h
