+++ include/con4m/datatypes.h
@@ -25,15 +25,15 @@ typedef struct hatrack_set_st c4m_set_t;
 #include "con4m/datatypes/streams.h"
 #include "con4m/datatypes/format.h"
 #include "con4m/datatypes/vm.h"
-#include "frontend/datatypes/lex.h"
-#include "frontend/datatypes/error.h"
-#include "frontend/datatypes/parse.h"
-#include "frontend/datatypes/scope.h"
-#include "frontend/datatypes/partials.h"
-#include "frontend/datatypes/spec.h"
-#include "frontend/datatypes/cfg.h"
-#include "frontend/datatypes/file.h"
-#include "frontend/datatypes/compile.h"
+#include "compiler/datatypes/lex.h"
+#include "compiler/datatypes/error.h"
+#include "compiler/datatypes/parse.h"
+#include "compiler/datatypes/scope.h"
+#include "compiler/datatypes/partials.h"
+#include "compiler/datatypes/spec.h"
+#include "compiler/datatypes/cfg.h"
+#include "compiler/datatypes/file.h"
+#include "compiler/datatypes/compile.h"
 
 typedef c4m_str_t *(*c4m_repr_fn)(c4m_obj_t, to_str_use_t);
 typedef void (*c4m_marshal_fn)(c4m_obj_t,
