+++ meson.build
@@ -93,19 +93,19 @@ c4m_src  = ['src/con4m/style.c',
             'src/con4m/fptostr.c',
             'src/con4m/path.c',
             'src/con4m/crypto/sha.c',
-            'src/con4m/frontend/compile.c',
-            'src/con4m/frontend/lex.c',
-            'src/con4m/frontend/parse.c',
-            'src/con4m/frontend/errors.c',
-            'src/con4m/frontend/scope.c',
-            'src/con4m/frontend/specs.c',
-            'src/con4m/frontend/cfg_build.c',            
-            'src/con4m/frontend/cfg.c',
-            'src/con4m/frontend/ast_utils.c',
-            'src/con4m/frontend/decl_pass.c',
-            'src/con4m/frontend/check_pass.c',
-            'src/con4m/frontend/memory_layout.c',
-            'src/con4m/frontend/codegen.c',
+            'src/con4m/compiler/compile.c',
+            'src/con4m/compiler/lex.c',
+            'src/con4m/compiler/parse.c',
+            'src/con4m/compiler/errors.c',
+            'src/con4m/compiler/scope.c',
+            'src/con4m/compiler/specs.c',
+            'src/con4m/compiler/cfg_build.c',            
+            'src/con4m/compiler/cfg.c',
+            'src/con4m/compiler/ast_utils.c',
+            'src/con4m/compiler/decl_pass.c',
+            'src/con4m/compiler/check_pass.c',
+            'src/con4m/compiler/memory_layout.c',
+            'src/con4m/compiler/codegen.c',
             ]
 
 hat_primary = ['src/hatrack/support/hatrack_common.c',
