+++ src/con4m/compiler/decl_pass.c
