+++ src/con4m/compiler/lex.c
