+++ src/con4m/compiler/scope.c
