+++ src/con4m/compiler/specs.c
